backend/server/requirements.txt
Django==5.2.8
djangorestframework
django-cors-headers
psycopg2-binary
gunicorn
dj-database-url
whitenoise
python-dotenv
djangorestframework-simplejwt
